test(static mediator): test dynamic and static mediators separately
